+++ jitlinkc++/rtdyld.h
@@ -24,7 +24,7 @@ extern "C"
   } SectionNameRTDYLD;
 
   typedef void *(*getfn_ptr)(void *, const char *, size_t);
-  typedef void (*offset_ptr)(void *, const char *, size_t, unsigned);
+  typedef void (*offset_ptr)(void *, const char *, size_t, uint64_t);
 
   typedef AllocBlockSliceRTDYLD (*allocate_t_rtdyld)(void *, AllocRequestRTDYLD, SectionNameRTDYLD);
